[CORDA-1778]: Moved ConfigValueExtractor to its right place.
