+++ node/src/main/kotlin/net/corda/node/services/config/parsing/Configuration.kt
@@ -1,6 +1,7 @@
 package net.corda.node.services.config.parsing
 
 import com.typesafe.config.Config
+import com.typesafe.config.ConfigException
 import com.typesafe.config.ConfigValue
 
 object Configuration {
@@ -10,8 +11,24 @@ object Configuration {
         fun describe(configuration: Config): ConfigValue
     }
 
-    interface ValueExtractor {
+    object Value {
 
+        interface Extractor<TYPE> {
+
+            @Throws(ConfigException.Missing::class, ConfigException.WrongType::class, ConfigException.BadValue::class)
+            fun valueIn(configuration: Config): TYPE
+
+            fun isSpecifiedBy(configuration: Config): Boolean
+
+            @Throws(ConfigException.WrongType::class, ConfigException.BadValue::class)
+            fun valueInOrNull(configuration: Config): TYPE? {
+
+                return when {
+                    isSpecifiedBy(configuration) -> valueIn(configuration)
+                    else -> null
+                }
+            }
+        }
     }
 
     object Property {
